Channel details cleanup. Not finalized.
