+++ contentcuration/contentcuration/frontend/channelList/views/Channel/ChannelModal.vue
@@ -67,32 +67,7 @@
                 />
               </VFlex>
             </VLayout>
-            <VLayout v-if="channelDetails && channelDetails.resource_count" row wrap justify-center>
-              <VFlex xs4>
-                <VMenu offset-y>
-                  <template v-slot:activator="{ on }">
-                    <VBtn
-                      color="primary"
-                      dark
-                      v-on="on"
-                    >
-                      {{ $tr('downloadReport') }}
-                    </VBtn>
-                  </template>
-                  <VList>
-                    <VListTile
-                      v-for="(item, index) in downloadOptions"
-                      :key="index"
-                      :href="item.href"
-                      download
-                    >
-                      <VListTileTitle>{{ item.title }}</VListTileTitle>
-                    </VListTile>
-                  </VList>
-                </VMenu>
-              </VFlex>
-            </VLayout>
-            <ChannelDetails  v-if="channelDetails && channelDetails.resource_count"/>
+            <ChannelDetails  v-if="channelDetails && channel.count" :channelId="channelId"/>
             <template v-else>
               {{ $tr('empty_details') }}
             </template>
@@ -100,6 +75,27 @@
         </VLayout>
       </VCardText>
       <VCardActions v-if="channel.edit">
+        <VMenu v-if="channel.count" offset-y>
+          <template v-slot:activator="{ on }">
+            <VBtn
+              class="upper"
+              color="info"
+              v-on="on"
+            >
+              {{ $tr('downloadReport') }}
+            </VBtn>
+          </template>
+          <VList>
+            <VListTile
+              v-for="(item, index) in downloadOptions"
+              :key="index"
+              :href="item.href"
+              download
+            >
+              <VListTileTitle>{{ item.title }}</VListTileTitle>
+            </VListTile>
+          </VList>
+        </VMenu>
         <VSpacer/>
         <VBtn class="upper" color="error" @click="deleteDialog=true">
           {{ $tr('deleteChannel') }}
@@ -142,32 +138,14 @@
 
   // Components
   import PrimaryDialog from 'shared/views/PrimaryDialog';
-  import CopyToken from 'shared/views/CopyToken';
   import ChannelStar from './ChannelStar';
   import ChannelDetails from './ChannelDetails';
   import LanguageDropdown from 'edit_channel/sharedComponents/LanguageDropdown';
   import ThumbnailUpload from 'shared/views/ThumbnailUpload';
 
-  const SCALE_TEXT = [
-    'very_small',
-    'very_small',
-    'small',
-    'small',
-    'average',
-    'average',
-    'average',
-    'large',
-    'large',
-    'very_large',
-    'very_large',
-  ];
-
-  const CHANNEL_SIZE_DIVISOR = 100000000;
-
   export default {
     name: 'ChannelModal',
     components: {
-      CopyToken,
       ChannelStar,
       ChannelDetails,
       LanguageDropdown,
@@ -197,64 +175,7 @@
       deletingChannel: 'Deleting channel...',
       deleteWarning:
         'All content under this channel will be deleted.\nAre you sure you want to delete this channel?',
-      authors: 'This channel features resources created by',
-      aggregators: 'Material in this channel was originally hosted at',
-      providers: 'The material in this channel was provided by',
       empty_details: 'This channel is empty',
-      topic_author: 'This topic features resources created by',
-      topic_aggregator: 'Material in this topic was originally hosted at',
-      topic_provider: 'The material in this topic was provided by',
-      topic_empty_details: 'This topic is empty',
-      no_license: 'No license selected',
-      author_description: 'Person or organization who created the content',
-      aggregator_description:
-        'Website or org hosting the content collection but not necessarily the creator or copyright holder',
-      provider_description: 'Organization that commissioned or is distributing the content',
-      license: '{count, plural,\n =1 {License}\n other {Licenses}}',
-      copyright_holder: '{count, plural,\n =1 {Copyright Holder}\n other {Copyright Holders}}',
-      auth_info: 'Authoring Information',
-      metadata_info: 'Content Metadata',
-      summary_info: 'Summary',
-      total_resources: '# of Resources',
-      resource_size: 'Size',
-      storage: 'Storage',
-      visibility_breakdown: 'Visibility',
-      content_breakdown: 'Content Summary',
-      languages: 'Languages',
-      tags: 'Content Tags',
-      resource_count: '{count, plural,\n =1 {# Resource}\n other {# Resources}}',
-      visibility_count:
-        '{count, plural,\n =1 {# resource is}\n other {# resources are}} visible to {user}',
-      kind_count: '{count, plural,\n =1 {# {kind}}\n other {# {kind_plural}}}',
-      role_description: 'Coach content is visible to coaches only in Kolibri',
-      sample_pathway: 'Sample Pathway',
-      channel_id: 'Channel ID',
-      channel_tokens: '{count, plural,\n =1 {Channel Token}\n other {Channel Tokens}}',
-      copy: 'Copy',
-      copy_text: 'Copy the following into Kolibri to import this channel',
-      total_resource_count: '{data, plural,\n =1 {Total Resource}\n other {Total Resources}}',
-      invalid_channel: 'Cannot save invalid channel',
-      original_channels: 'Includes Content From',
-      whats_inside: "What's Inside",
-      source: 'Source',
-      topic: 'topic',
-      using_channel: 'Using this Channel',
-      very_small: 'Very Small',
-      small: 'Small',
-      average: 'Average',
-      large: 'Large',
-      very_large: 'Very Large',
-      includes: 'Includes',
-      coach_content: 'Coach Content',
-      assessments: 'Assessments',
-      accessible_languages: 'Subtitles',
-      instructor_resources: 'For Educators',
-      recommended: '(Recommended)',
-      preview: 'Preview',
-      more: 'Show More ({more})',
-      less: 'Show Less',
-      original_content: 'Original Content',
-      details_tooltip: '{kind} ({percent}%)',
       downloadReport: 'Download Channel Report',
       downloadDetailedPDF: 'Download Detailed PDF',
       downloadPDF: 'Download PDF',
@@ -366,39 +287,6 @@
           }
         },
       },
-      countBar() {
-        // Get data for count bar indicator
-        if (this.channelDetails && this.channelDetails.resource_count) {
-          const sizeIndex = Math.max(
-            1,
-            Math.min(Math.floor(Math.log(this.channelDetails.resource_count) / Math.log(2.8)), 10)
-          );
-          const bar = [];
-          for (var i = 0; i < 10; ++i) {
-            bar.push(i < sizeIndex);
-          }
-          return {
-            filled: bar,
-            text: this.$tr(SCALE_TEXT[sizeIndex]),
-          };
-        }
-      },
-      sizeBar() {
-        // Get data for count bar indicator
-        if (this.channelDetails && this.channelDetails.size) {
-          const sizeIndex = Math.max(
-            1,
-            Math.min(
-              Math.ceil(Math.log(this.channelDetails.size / CHANNEL_SIZE_DIVISOR) / Math.log(2)),
-              10
-            )
-          );
-          return {
-            filled: Array(sizeIndex),
-            text: this.$tr(SCALE_TEXT[sizeIndex]),
-          };
-        }
-      },
       downloadOptions() {
         return [
           {
@@ -445,12 +333,6 @@
             reject();
           })
         });
-      },
-      channelLink(channelId) {
-        return window.Urls.channel() + `#/${channelId}/view/`;
-      },
-      nodeLink(nodeId) {
-        return this.channelLink(this.channelId) + nodeId;
       },
       setChannelDetails(channelId) {
         if (!isTempId(channelId) && !this.channelDetails) {
