+++ package.json
@@ -100,6 +100,7 @@
     "vue": "^2.5.17",
     "vue-intl": "^3.0.0",
     "vuetify": "^1.5.16",
+    "vuewordcloud": "^18.7.11",
     "vuex": "^3.0.1"
   },
   "devDependencies": {
